+++ src/instruments/src/EFB/Assets/Theme.css
@@ -2,7 +2,7 @@
 :root.theme-blue{
     --color-highlight: #00c9e4;
     --color-body: #0E131B;
-    --color-secondary: #171E2C;
+    --color-secondary: #1d2129;
     --color-text: #FFFFFF;
     --color-unselected: #C9C9C9;
     --color-accent: #2b313b;
